move AvroSerializer.java to src; make EnvProperties extension
