+++ backend/kafka-producer/src/main/java/com/kafka/viewer/config/ProducerProperty.java
@@ -2,17 +2,7 @@ package com.kafka.viewer.config;
 
 public class ProducerProperty {
     /**
-     * Kafka bootstrap server
-     */
-    public static final String KAFKA_BOOTSTRAP_SERVER = "kafka.bootstrap.server";
-
-    /**
      * Kafka topic for publish
      */
     public static final String KAFKA_TOPIC = "kafka.topic";
-
-    /**
-     * Generator class
-     */
-    public static final String GENERATOR_CLASS = "generator.class";
 }
