+++ backend/kafka-producer/src/main/java/com/kafka/viewer/producer/AvroSerializer.java
