+++ backend/kafka-producer/src/main/resources/application.properties
@@ -1,2 +1,6 @@
-kafka.bootstrap.server=${KAFKA_BOOTSTRAP_SERVER:localhost:9092}
-kafka.topic=${KAFKA_TOPIC}
+# Bootstrap server
+bootstrap.servers=${KAFKA_BOOTSTRAP_SERVER:localhost:9092}
+
+# Value serializer
+key.serializer=org.apache.kafka.common.serialization.LongSerializer
+value.serializer=com.kafka.viewer.producer.AvroSerializer
