+++ backend/kafka-producer/src/test/java/com/kafka/viewer/KafkaProducerApplicationTest.java
@@ -6,12 +6,5 @@ import org.junit.jupiter.api.Test;
 
 /**
  */
-public class KafkaProducerApplicationTest 
-{
-    /**
-     */
-    @Test
-    @DisplayName("")
-    public void should() {
-    }
+public class KafkaProducerApplicationTest {
 }
