+++ backend/kafka-producer/src/test/resources/application.properties
@@ -1,2 +1,7 @@
-kafka.bootstrap.server=${KAFKA_BOOTSTRAP_SERVER:localhost:9092}
+# Bootstrap server
+bootstrap.servers=${KAFKA_BOOTSTRAP_SERVER:localhost:9092}
 kafka.topic=${KAFKA_TOPIC:test-topic}
+
+# Value serializer
+key.serializer=org.apache.kafka.common.serialization.LongSerializer
+value.serializer=com.kafka.viewer.producer.AvroSerializer
